requirements.txt
streamlit>=1.33.0
streamlit-folium>=0.18.0
folium>=0.15.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.5.0
lxml>=5.1.0
python-dateutil>=2.8.0
feedparser>=6.0.0
